+++ tests/test_database.py
@@ -159,7 +159,7 @@ def test_db_metadata_path(db_fixt):
 
 
 def test_db_land_path(db_fixt):
-    from tests.helpers import land_bucket
+    from tests.conftest import land_bucket
 
     db, _ = db_fixt
     exp_lp = os.path.join("s3://", land_bucket, "dep", db.env, db.name)
@@ -167,7 +167,7 @@ def test_db_land_path(db_fixt):
 
 
 def test_db_raw_hist_path(db_fixt):
-    from tests.helpers import raw_hist_bucket
+    from tests.conftest import raw_hist_bucket
 
     db, _ = db_fixt
     exp_rhp = os.path.join("s3://", raw_hist_bucket, "dep", db.env, db.name)
@@ -180,7 +180,7 @@ def test_db_processed_path(db_fixt):
 
 
 def test_db_curated_path(db_fixt):
-    from tests.helpers import dep_bucket
+    from tests.conftest import dep_bucket
 
     db, _ = db_fixt
     exp_cp = os.path.join("s3://", dep_bucket, db.env, db.name, "curated")
