--- tests/test_pipeline_builder.py
@@ -1,9 +1,0 @@
-import toml
-
-from src import __version__
-
-
-def test_pyproject_toml_matches_version():
-    with open("pyproject.toml") as f:
-        proj = toml.load(f)
-    assert __version__ == proj["tool"]["poetry"]["version"]
